Hackwekend/GraphQL/Topic2/requirements.txt
flask
ariadne
uvicorn
python-dotenv
tinydb
requests